A bit more progress/reorganization
