+++ server/paper/src/main/kotlin/net/stellarica/server/material/custom/item/CustomItem.kt
@@ -1,14 +1,8 @@
 package net.stellarica.server.material.custom.item
 
-import net.kyori.adventure.text.Component
 import net.minecraft.resources.ResourceLocation
-import net.stellarica.server.StellaricaServer.Companion.plugin
 import net.stellarica.server.material.type.item.VanillaItemType
-import net.stellarica.server.utils.extensions.asMiniMessage
-import org.bukkit.NamespacedKey
 import org.bukkit.enchantments.Enchantment
-import org.bukkit.inventory.ItemStack
-import org.bukkit.persistence.PersistentDataType
 
 /**
  * Holds data for a custom item
@@ -17,53 +11,18 @@ data class CustomItem(
 	/** The ID of the item. */
 	val id: ResourceLocation,
 
-	/**
-	 * The display name of the item, MiniMessage formatting is allowed
-	 */
+	/** The display name of the item, MiniMessage formatting is allowed */
 	val name: String,
 
-	/**
-	 * The item's lore, MiniMessage formatting is allowed
-	 */
+	/** The item's lore, MiniMessage formatting is allowed */
 	val lore: List<String>,
 
 	/** The vanilla material behind this item */
 	val base: VanillaItemType,
 
-	/**
-	 * Custom model data for this item
-	 */
+	/** Custom model data for this item */
 	val modelData: Int,
 
 	/** The allowed enchants for this item */
 	val allowedEnchants: MutableSet<Enchantment>? = null
-) {
-
-	/**
-	 * @return an [ItemStack] with [count] of this item
-	 */
-	fun getItem(count: Int = 1): ItemStack {
-		val stack = ItemStack(base, count)
-		stack.editMeta { meta ->
-			meta.displayName(name.asMiniMessage)
-			val loreComponents = mutableListOf<Component>() // this can be code golfed
-			lore.forEach { loreComponents.add(it.asMiniMessage) }
-			if (maxPower > 0) {
-				loreComponents.add("<gray>Power: 0/$maxPower".asMiniMessage)
-				meta.persistentDataContainer.set(
-					NamespacedKey(plugin, "power"),
-					PersistentDataType.INTEGER,
-					0
-				)
-			}
-			meta.lore(loreComponents)
-			meta.setCustomModelData(modelData)
-			meta.persistentDataContainer.set(
-				NamespacedKey(plugin, "custom_item_id"),
-				PersistentDataType.STRING,
-				id
-			)
-		}
-		return stack
-	}
-}
+)
