+++ server/paper/src/main/kotlin/net/stellarica/server/material/type/block/CustomBlockType.kt
@@ -1,8 +1,8 @@
-package net.stellarica.server.material.block
+package net.stellarica.server.material.type.block
 
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.level.block.Block
-import net.stellarica.server.customblocks.CustomBlock
+import net.stellarica.server.material.custom.block.CustomBlock
 import org.bukkit.Material
 import org.bukkit.block.data.BlockData
 
@@ -27,9 +27,4 @@ value class CustomBlockType(val type: CustomBlock) : BlockType {
 	override fun getId(): ResourceLocation {
 		TODO("Not yet implemented")
 	}
-
-	override fun getStringId(): String {
-		TODO("Not yet implemented")
-	}
-
 }
